refactor: update type definitions and reorganize audio pool structure for improved clarity
