+++ src/types/pool.ts
@@ -1,0 +1,9 @@
+export interface Pool {
+  audio: HTMLAudioElement
+  source: string
+  isPlaying: boolean
+  cleanupListeners: (() => void)[]
+  onEnd?: () => void
+  lastUsed: number
+  duration?: number
+}
